requirements.txt
asgiref==3.1.4
asn1crypto==0.24.0
attrs==19.1.0
autobahn==19.7.2
Automat==0.7.0
certifi==2019.6.16
cffi==1.12.3
channels==2.2.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.7
daphne==2.3.0
defusedxml==0.6.0
Django==2.2.3
django-allauth==0.39.1
django-cors-headers==3.0.2
django-filters==0.2.1
django-templated-mail==1.1.1
django-webpack-loader==0.6.0
djangorestframework==3.10.1
djangorestframework-simplejwt==4.3.0
djoser==2.0.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
oauthlib==3.0.2
Pillow==6.1.0
psycopg2==2.8.3
pycparser==2.19
PyHamcrest==1.9.0
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.1
pywin32==224
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
Twisted==19.2.1
txaio==18.8.1
urllib3==1.25.3
whitenoise==4.1.3
zope.interface==4.6.0
